requirements.txt
# Production dependencies
PyYAML>=6.0.1

# Development dependencies
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional development tools
mypy>=1.5.0
coverage>=7.3.0
